Refactor: Course Service to include a top level interface Added robust exception handling to prevent duplicate course creations
